Split main function from ulp initialization
