+++ src/main.c
@@ -1,0 +1,26 @@
+#include <stdio.h>
+#include <string.h>
+#include "esp_sleep.h"
+#include "ulp_adc_example.h"
+
+
+
+void app_main()
+{
+    esp_sleep_wakeup_cause_t cause = esp_sleep_get_wakeup_cause();
+    if (cause != ESP_SLEEP_WAKEUP_ULP) {
+        printf("Not ULP wakeup\n");
+        init_ulp_program();
+    } else {
+        printf("Deep sleep wakeup\n");
+        printf("ULP did %d measurements since last reset\n", ulp_sample_counter & UINT16_MAX);
+        printf("Thresholds:  low=%d  high=%d\n", ulp_low_thr, ulp_high_thr);
+        ulp_last_result &= UINT16_MAX;
+        printf("Value=%d was %s threshold\n", ulp_last_result,
+                ulp_last_result < ulp_low_thr ? "below" : "above");
+    }
+    printf("Entering deep sleep\n\n");
+    start_ulp_program();
+    ESP_ERROR_CHECK( esp_sleep_enable_ulp_wakeup() );
+    esp_deep_sleep_start();
+}
